Made FieldWithIcon a reusable component throughout the application
