+++ src/Components/DataEntry/FieldWithIcon/FieldWithIcon.css
@@ -1,0 +1,37 @@
+label {
+  margin-top: 50px;
+  margin-left: 2vh;
+  color: #626262;
+  font-weight: 500;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+input {
+  margin-top: 5px;
+  border: 1px solid #ddd;
+  border-radius: 50vh;
+  height: 48px;
+  padding: 0 40px 0 50px;
+  font-size: 16px;
+  color: #666;
+  width: 100%;
+
+  box-sizing: border-box;
+}
+
+input::placeholder {
+  color: #999;
+}
+
+.input-icon-container {
+  position: relative;
+}
+
+.input-icon {
+  position: absolute;
+  color: #bbb;
+  min-width: 50px;
+  text-align: center;
+  top: calc(50% - 6px);
+  left: 5px;
+}
