+++ src/Views/Login/Login.css
@@ -12,44 +12,6 @@
   flex-direction: column;
 }
 
-.login-container form label {
-  margin-top: 50px;
-  margin-left: 2vh;
-  color: #626262;
-  font-weight: 500;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.input-icon-container {
-  position: relative;
-}
-
-.input-icon {
-  position: absolute;
-  color: #bbb;
-  min-width: 50px;
-  text-align: center;
-  top: calc(50% - 5px);
-  left: 5px;
-}
-
-.login-container form input {
-  margin-top: 5px;
-  border: 1px solid #ddd;
-  border-radius: 50vh;
-  height: 48px;
-  padding: 0 40px 0 50px;
-  font-size: 16px;
-  color: #666;
-  width: 100%;
-
-  box-sizing: border-box;
-}
-
-.login-container form input::placeholder {
-  color: #999;
-}
-
 .login-container form .authentication-error {
   color: #eb4d4b;
   margin-top: 10px;
